Changed name to be more fitting of actual purpose
Old name was literally a joke based on the initial idea that spawned the bot ("What if we hooked up a power supply to tase me when people input a Discord command?" - my friend).  It's outdated.  While this bot is not done, and might never be done (as I have no time to complete it now), this is still a pretty good portfolio piece for me now so having an accurate, up-to-date name helps me there.